refactor: move node manager into dedicated module
